pyproject.toml
[project]
name = "kejebe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.2",
    "django-filter>=25.1",
    "django-mptt>=0.17.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
]
